+++ app/code/community/Ffuenf/Pagespeed/Test/Config/Main.php
@@ -1,0 +1,100 @@
+<?php
+/**
+ * Ffuenf_Pagespeed extension.
+ *
+ * NOTICE OF LICENSE
+ *
+ * This source file is subject to the MIT License
+ * that is bundled with this package in the file LICENSE.txt.
+ * It is also available through the world-wide-web at this URL:
+ * http://opensource.org/licenses/mit-license.php
+ *
+ * @category   Ffuenf
+ *
+ * @author     Achim Rosenhagen <a.rosenhagen@ffuenf.de>
+ * @copyright  Copyright (c) 2015 ffuenf (http://www.ffuenf.de)
+ * @license    http://opensource.org/licenses/mit-license.php MIT License
+ */
+
+/**
+ * @loadSharedFixture shared
+ */
+class Ffuenf_Pagespeed_Test_Config_Main extends EcomDev_PHPUnit_Test_Case_Config
+{
+    /**
+     * Check if the installed module has the correct module version
+     *
+     * @test
+     */
+    public function testModuleConfig()
+    {
+        $this->assertModuleVersionGreaterThanOrEquals($this->expected('module')->getVersion(), 'module is new enough');
+        $this->assertModuleCodePool($this->expected('module')->getCodePool(), 'correct module code pool');
+        $this->assertModuleIsActive('module is active');
+    }
+
+    /**
+     * Tests whether extension helper aliases are returning the correct class names
+     *
+     * @test
+     */
+    public function testHelperAliases()
+    {
+        $this->assertHelperAlias(
+            'pagespeed',
+            'Ffuenf_Pagespeed_Helper_Data'
+        );
+    }
+
+    /**
+     * Tests whether extension uses the old-style admin routing (not compatible with SUPEE-6788).
+     *
+     * @test
+     */
+    public function testGetOldAdminRouting()
+    {
+        $routers = Mage::getConfig()->getNode('admin/routers');
+        $offendingExtensions = array();
+        foreach ($routers[0] as $router) {
+            $name = $router->args->module;
+            if ($name != 'Mage_Adminhtml') {
+                $offendingExtensions[] = $router->args->module;
+            }
+        }
+        $this->assertEquals(
+            count($offendingExtensions),
+            0,
+            'This extension uses old-style admin routing which is not compatible with SUPEE-6788 / Magento 1.9.2.2+'
+        );
+    }
+
+    /**
+     * Tests filecachepath.
+     *
+     * @test
+     * @covers Ffuenf_Pagespeed_Helper_Data::GetFilecachepath
+     */
+    public function testGetFilecachepath()
+    {
+        $this->assertEquals(
+            Mage::helper('pagespeed')->getFilecachepath(),
+            '/var/ngx_pagespeed_cache',
+            'Filecachpath is not set please check config'
+        );
+    }
+
+    /**
+     * Tests cacheflushfilename.
+     *
+     * @test
+     * @covers Ffuenf_Pagespeed_Helper_Data::GetFilecachepath
+     */
+    public function testGetCacheflushfilename()
+    {
+        $this->assertEquals(
+            Mage::helper('pagespeed')->getCacheflushfilename(),
+            'cache.flush',
+            'Cacheflushfilename is not set please check config'
+        );
+    }
+}